requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==1.5.0
azure-functions-durable>=1.0.0b12
pandas
pyodbc==4.0.30